rotas de receitas completa e ajustes no crud

+++ recita-facil-site/routes/web.php
@@ -20,17 +20,21 @@ route::get('/receitas',[ReceitasController::class, 'index'])->name('receitas.ind
 
 // ROTAS DE USUÁRIO
 Route::get('/users',[UserController::class, 'index'])->middleware(['auth','is_admin'])->name('users.index');
-Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
-Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update');
-Route::post('/users', [UserController::class, 'store'])->name('users.store');
-Route::delete('/users/{id}',[UserController::class, 'delete'])->name('users.delete');
-Route::get('/users/{id}',[UserController::class,'show'])->name('users.show');
+Route::get('/users/create', [UserController::class, 'create'])->middleware(['auth','is_admin'])->name('users.create');
+Route::get('/users/{id}/edit', [UserController::class, 'edit'])->middleware(['auth','is_admin'])->name('users.edit');
+Route::put('/users/{id}', [UserController::class, 'update'])->middleware(['auth','is_admin'])->name('users.update');
+Route::post('/users', [UserController::class, 'store'])->middleware(['auth','is_admin'])->name('users.store');
+Route::delete('/users/{id}',[UserController::class, 'delete'])->middleware(['auth','is_admin'])->name('users.delete');
+Route::get('/users/{id}',[UserController::class,'show'])->middleware(['auth','is_admin'])->name('users.show');
 
 // ROTAS DE RECEITAS
 Route::get('/receitas', [RecipesController::class, 'index'])->name('recipes.index');
-Route::get('/receitas/create', [RecipesController::class, 'create'])->name('recipes.create');
-Route::post('/receitas', [RecipesController::class, 'store'])->name('recipes.store');
+Route::get('/receitas/create', [RecipesController::class, 'create'])->middleware(['auth','is_admin'])->name('recipes.create');
+Route::get('/receitas/{id}/edit', [RecipesController::class, 'edit'])->name('recipes.edit');
+Route::put('recipes/{id}', [RecipesController::class, 'update'])->name('recipes.update');
+Route::post('/receitas', [RecipesController::class, 'store'])->middleware(['auth','is_admin'])->name('recipes.store');
+Route::delete('/receitas/{id}', [RecipesController::class, 'delete'])->name('recipes.delete');
+Route::get('/recipes/{id}', [RecipesController::class, 'show'])->name('recipes.show');
 
 
 Route::get('/dashboard', function () {
